clean up frontend and make adding a rom one step
